+++ libs/cardano-ledger-core/src/Cardano/Ledger/BaseTypes.hs
@@ -60,7 +60,7 @@ module Cardano.Ledger.BaseTypes (
   dnsToText,
   textToDns,
   Port (..),
-  Milliseconds (..),
+  Milliseconds32 (..),
   ActiveSlotCoeff,
   mkActiveSlotCoeff,
   activeSlotVal,
@@ -703,7 +703,7 @@ newtype Port = Port {portToWord16 :: Word16}
 
 -- | A non-negative duration in milliseconds using 32 bits, yielding a
 -- `maxBound` of ~49.7 days.
-newtype Milliseconds = Milliseconds {unMilliseconds :: Word32}
+newtype Milliseconds32 = Milliseconds32 {unMilliseconds32 :: Word32}
   deriving (Eq, Ord, Generic, Show)
   deriving newtype (Num, Bounded, DecCBOR, EncCBOR, NFData, NoThunks, ToJSON, FromJSON)